+ Se agregan validaciones al metodo de captura de eventos de mouse. + Se agrega funcionalidad a la grilla y se prueba minimamente en el modo DDA

+++ tp1/Geometria/Segmento.cpp
@@ -153,7 +153,7 @@ void Segmento::rellenar() {
 
 }
 
-bool Segmento::contiene(float x, float y){
+bool Segmento::contiene(int x, int y){
     //TODO
     return false;
 }
